Bug 1706866 - Change radix type to uint8_t in JS::SimpleStringToBigInt(). r=jandem
There are actually several places in vm/BigIntType.h where unsigned is
used for the radix, and several other places where uint8_t is used. In the
public API, prefer the narrower type, for consistency with the API that
will be added in the next commit.

https://bugzilla.mozilla.org/show_bug.cgi?id=1706866

Differential Revision: https://phabricator.services.mozilla.com/D117191

UltraBlame original commit: c2ace7587c0d222e617ab76e248346ce94bbaa1e

+++ js/public/BigInt.h
@@ -113,7 +113,7 @@ extern JS_PUBLIC_API BigInt* StringToBigInt(
 
 
 extern JS_PUBLIC_API BigInt* SimpleStringToBigInt(
-    JSContext* cx, mozilla::Span<const char> chars, unsigned radix);
+    JSContext* cx, mozilla::Span<const char> chars, uint8_t radix);
